continue work with logic fix some bugs, add some code, rewrite some queries from jpa repo to mongo etc...

+++ src/main/java/com/certTrack/CourseService/Security/JWTToPrincipalConverter.java
@@ -15,7 +15,7 @@ public class JWTToPrincipalConverter {
 	
 	public UserPrincipal convert(DecodedJWT jwt) {
 		return UserPrincipal.builder()
-				.userId(Long.valueOf(jwt.getSubject()))
+				.userId(Integer.valueOf(jwt.getSubject()))
 				.email(jwt.getClaim("e").asString())
 				.authorities(extractAuthoritiesFromClaim(jwt))
 				.build();
